refactor(seo): Update all the pages into static pages by separating it to pages component
